+++ src/ts/PlotBoilerplate.ts
@@ -75,7 +75,8 @@
  * @modified 2021-04-25 Extending `remove` to accept arrays of drawables.
  * @modified 2021-11-16 Adding the `PBText` drawable.
  * @modified 2022-08-01 Added `title` to the params.
- * @version  1.15.1
+ * @modified 2022-10-25 Added the `origin` to the default draw config.
+ * @version  1.16.0
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -532,6 +533,9 @@ export class PlotBoilerplate {
         lineWidth: 1,
         fill: true,
         anchor: true
+      },
+      origin: {
+        color: "#000000"
       }
     }; // END drawConfig
 
@@ -1095,7 +1099,7 @@ export class PlotBoilerplate {
   drawOrigin(draw: DrawLib<any>) {
     // Add a crosshair to mark the origin
     draw.setCurrentId("origin");
-    draw.crosshair({ x: 0, y: 0 }, 10, "#000000");
+    draw.crosshair({ x: 0, y: 0 }, 10, this.drawConfig.origin.color);
   }
 
   /**