+++ src/ts/Vector.ts
@@ -11,7 +11,8 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `Vector.toSVGString` function (deprecated). Use `drawutilssvg` instead.
- * @version  1.4.0
+ * @modified 2022-10-25 Added the `getOrthogonal` method.
+ * @version  1.5.0
  *
  * @file Vector
  * @public
@@ -65,7 +66,7 @@ export class Vector extends VertTuple<Vector> implements SVGSerializable {
   }
 
   /**
-   * The inverse of a vector is a vector witht the same magnitude but oppose direction.
+   * The inverse of a vector is a vector with the same magnitude but oppose direction.
    *
    * Please not that the origin of this vector changes here: a->b becomes b->a.
    *
@@ -114,6 +115,25 @@ export class Vector extends VertTuple<Vector> implements SVGSerializable {
 
     // if we cast these lines infinitely in both directions, they intersect here:
     return new Vertex(this.a.x + a * (this.b.x - this.a.x), this.a.y + a * (this.b.y - this.a.y));
+  }
+
+  /**
+   * Get the orthogonal "vector" of this vector (rotated by 90° clockwise).
+   *
+   * @name getOrthogonal
+   * @method getOrthogonal
+   * @return {Vector} A new vector with the same length that stands on this vector's point a.
+   * @instance
+   * @memberof Vector
+   **/
+  getOrthogonal(): Vector {
+    // Orthogonal of vector (0,0)->(x,y) is (0,0)->(-y,x)
+    const linePoint = this.a.clone();
+    const startPoint = this.b.clone().sub(this.a);
+    const tmp = startPoint.x;
+    startPoint.x = -startPoint.y;
+    startPoint.y = tmp;
+    return new Vector(linePoint, startPoint.add(this.a));
   }
 
   static utils = {